+++ web/src/locales/en.json
@@ -98,6 +98,12 @@
 	"signOut.button": "Sign out",
 	"myInformation": "My information",
 	"dashboard": "Dashboard",
+	"dashboard.activeEmployees": "Active employees",
+	"dashboard.amountOfWarehouses": "Number of warehouses",
+	"dashboard.amountOfTrucks": "Number of trucks",
+	"dashboard.containersAdded": "Containers added",
+	"dashboard.containersByCategory": "Containers by category",
+	"dashboard.containersByMunicipality": "Containers by municipality",
 	"map": "Map",
 	"departure": "Departure",
 	"arrival": "Arrival",
