+++ web/src/locales/pt.json
@@ -98,6 +98,12 @@
 	"signOut.button": "Terminar sessão",
 	"myInformation": "Minha informação",
 	"dashboard": "Painel",
+	"dashboard.activeEmployees": "Colaboradores ativos",
+	"dashboard.amountOfWarehouses": "Número de armazéns",
+	"dashboard.amountOfTrucks": "Número de camiões",
+	"dashboard.containersAdded": "Contentores adicionados",
+	"dashboard.containersByCategory": "Contentores por categoria",
+	"dashboard.containersByMunicipality": "Contentores por município",
 	"map": "Mapa",
 	"departure": "Partida",
 	"arrival": "Chegada",
