+++ web/src/routes/backOffice/dashboard/partials/ContainersAdded.svelte
@@ -4,30 +4,48 @@
 	import type { Container } from "../../../../domain/container";
 	import Spinner from "../../../../lib/components/Spinner.svelte";
 	import { getCssVariable } from "../../../../lib/utils/cssVars";
+	import { get } from "svelte/store";
+	import { locale, t } from "../../../../lib/utils/i8n";
+	import { getColorWithOpacity } from "../../../../lib/utils/color";
 
+	/**
+	 * The promise with the containers.
+	 */
 	export let containersPromise: Promise<Container[]>;
 
-	let loading = false;
+	/**
+	 * Indicates whether the containers are being loaded.
+	 */
+	let loading = true;
 
+	/**
+	 * The canvas element where the chart is rendered.
+	 */
+	let canvas: HTMLCanvasElement;
+
+	/**
+	 * Retrieves a map with the month index and the corresponding names.
+	 * @returns Map with the month index and the corresponding name.
+	 */
 	function getMonths(): Map<number, string> {
 		const today = new Date();
 		const monthsNamesMap = new Map<number, string>();
 
 		for (let i = 0; i < 12; i++) {
 			const date = new Date(today.getFullYear(), i, 1);
-			const month = date.toLocaleString("default", { month: "short" });
+			const month = date.toLocaleString(get(locale), { month: "short" });
 			monthsNamesMap.set(i, month);
 		}
 
 		return monthsNamesMap;
 	}
 
-	function makeChart(containers: Container[]) {
-		const canvasElement = document.getElementById("container-added-chart");
-		if (!(canvasElement instanceof HTMLCanvasElement)) {
-			return;
-		}
-
+	/**
+	 * Builds a chart with the containers.
+	 * @param containers Containers.
+	 */
+	function buildChart(containers: Container[]) {
+		// Build a map with the amount of containers added per month.
 		const containersPerMonth = new Map<number, number>();
 		const monthNames = getMonths();
 		for (const container of containers) {
@@ -42,6 +60,7 @@
 			containersPerMonth.set(monthDateInMilliseconds, amount + 1);
 		}
 
+		// Get chart labels.
 		const labels = Array.from(containersPerMonth.keys()).map(dateMs => {
 			const date = new Date(dateMs);
 			const monthName = monthNames.get(date.getMonth())!;
@@ -50,6 +69,7 @@
 			return `${monthName} ${year}`;
 		});
 
+		// Get chart data.
 		let data = Array.from(containersPerMonth.values());
 		data = data.reduce((acc, amount, idx) => {
 			if (idx > 0) {
@@ -59,7 +79,7 @@
 			return acc;
 		}, data);
 
-		new Chart(canvasElement, {
+		new Chart(canvas, {
 			type: "line",
 			data: {
 				labels,
@@ -67,6 +87,11 @@
 					{
 						data,
 						borderColor: getCssVariable("--green-700"),
+						backgroundColor: getColorWithOpacity(
+							getCssVariable("--green-700"),
+							0.2,
+						),
+						fill: "origin",
 					},
 				],
 			},
@@ -87,16 +112,16 @@
 		loading = false;
 	}
 
-	containersPromise.then(containers => makeChart(containers));
+	containersPromise.then(containers => buildChart(containers));
 </script>
 
 <Card element="article" class="containers-added-card">
-	<h2>Contentores adicionados</h2>
-	<canvas id="container-added-chart" style={loading ? "display: none;" : ""} />
+	<h2>{$t("dashboard.containersAdded")}</h2>
+	<canvas bind:this={canvas} style:display={loading ? "none" : ""} />
 	{#await containersPromise}
 		<Spinner />
 	{:catch}
-		<p>Erro</p>
+		<p>{$t("error.unexpected.title")}</p>
 	{/await}
 </Card>
