+++ web/src/routes/backOffice/dashboard/partials/ContainersByCategory.svelte
@@ -1,8 +1,6 @@
 <script lang="ts">
 	import { Chart } from "chart.js";
 	import Card from "../../components/Card.svelte";
-	import { getBatchPaginatedResponse } from "../../../../lib/utils/request";
-	import ecomapHttpClient from "../../../../lib/clients/ecomap/http";
 	import type {
 		Container,
 		ContainerCategory,
@@ -11,10 +9,24 @@
 	import { getCssVariable } from "../../../../lib/utils/cssVars";
 	import { t } from "../../../../lib/utils/i8n";
 
+	/**
+	 * The promise with the containers.
+	 */
 	export let containersPromise: Promise<Container[]>;
 
-	let loading = false;
+	/**
+	 * Indicates whether the containers are being loaded.
+	 */
+	let loading = true;
 
+	/**
+	 * The canvas element where the chart is rendered.
+	 */
+	let canvas: HTMLCanvasElement;
+
+	/**
+	 * The color for each respective container category.
+	 */
 	const categoryColors: Record<ContainerCategory, string> = {
 		general: getCssVariable("--sky-300"),
 		glass: getCssVariable("--lime-600"),
@@ -25,31 +37,41 @@
 		plastic: getCssVariable("--yellow-400"),
 	};
 
-	function makeChart(containers: Container[]) {
-		const canvasElement = document.getElementById("container-category-chart");
-		if (!(canvasElement instanceof HTMLCanvasElement)) {
-			return;
-		}
-
-		const data = new Map<ContainerCategory, number>();
+	/**
+	 * Builds a chart with the containers.
+	 * @param containers Containers.
+	 */
+	function buildChart(containers: Container[]) {
+		// Build a map with the amount of containers per container category.
+		const containerAmountPerCategory = new Map<ContainerCategory, number>();
 		for (const container of containers) {
-			const containerCategoryAmount = data.get(container.category) ?? 0;
-			data.set(container.category, containerCategoryAmount + 1);
+			const amount = containerAmountPerCategory.get(container.category) ?? 0;
+			containerAmountPerCategory.set(container.category, amount + 1);
 		}
 
-		const backgroundColor = Array.from(data.keys()).map(
+		// Get container categories.
+		const containerCategories = Array.from(containerAmountPerCategory.keys());
+
+		// Get chart labels.
+		const labels = containerCategories.map(category =>
+			$t(`containers.category.${category}`),
+		);
+
+		// Get chart data.
+		const data = Array.from(containerAmountPerCategory.values());
+
+		// Get background color for each container category.
+		const backgroundColor = containerCategories.map(
 			category => categoryColors[category],
 		);
 
-		new Chart(canvasElement, {
+		new Chart(canvas, {
 			type: "doughnut",
 			data: {
-				labels: Array.from(data.keys()).map(category =>
-					$t(`containers.category.${category}`),
-				),
+				labels,
 				datasets: [
 					{
-						data: Array.from(data.values()),
+						data,
 						backgroundColor,
 					},
 				],
@@ -66,19 +88,16 @@
 		loading = false;
 	}
 
-	containersPromise.then(containers => makeChart(containers));
+	containersPromise.then(containers => buildChart(containers));
 </script>
 
 <Card element="article" class="containers-category-card">
-	<h2>Contentores por categoria</h2>
-	<canvas
-		id="container-category-chart"
-		style={loading ? "display: none;" : ""}
-	/>
+	<h2>{$t("dashboard.containersByCategory")}</h2>
+	<canvas bind:this={canvas} style:display={loading ? "none" : ""} />
 	{#await containersPromise}
 		<Spinner />
-	{:catch error}
-		<p>{error}</p>
+	{:catch}
+		<p>{$t("error.unexpected.title")}</p>
 	{/await}
 </Card>
